refactor(navigation-header): simplify action group module
