--- src/frontend/components/headers/components/HeaderActionGroup/HeaderActionGroup.types.ts
@@ -1,7 +1,0 @@
-import type { HeaderActionTone, HeaderToolbarAction } from '../HeaderAction';
-
-export type HeaderActionGroupProps = {
-  actions: readonly HeaderToolbarAction[];
-  placement: 'left' | 'right';
-  tone?: HeaderActionTone;
-};
