--- src/frontend/components/headers/components/HeaderActionGroup/index.ts
@@ -1,2 +1,0 @@
-export { HeaderActionGroup } from './HeaderActionGroup';
-export type { HeaderActionGroupProps } from './HeaderActionGroup.types';
