+++ src/frontend/components/headers/index.ts
@@ -1,6 +1,9 @@
 export type { HeaderToolbarAction } from './components/HeaderAction';
 export { HeaderIconButton } from './components/HeaderAction';
-export { HeaderActionGroup, type HeaderActionGroupProps } from './components/HeaderActionGroup';
+export {
+  HeaderActionGroup,
+  type HeaderActionGroupProps,
+} from './components/HeaderActionGroup/HeaderActionGroup';
 export { HeaderChrome } from './components/HeaderChrome';
 export { headerScreenOptions } from './headerScreenOptions';
 export { MainHeader } from './MainHeader/MainHeader';
